+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v16.1.1...HEAD)
+* [PCM-2223](https://inindca.atlassian.net/browse/PCM-2223) Fix unintentional breaking change with types
 
 # [v16.1.1](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v16.1.0...v16.1.1)
 ### Added